Fix obsolete @tailwindcss/vite plugin
We encountered this error:

Error [ERR_MODULE_NOT_FOUND]: Cannot find package '@tailwindcss/vite' imported from ...

This plugin was previously used in some Laravel Breeze setup to
integrate TailwindCSS with Vite, but it's now deprecated and no longer
needed.  Using it causes issues like the above error.

To fix this error, make sure we have the correct dev dependencies:

$ sail npm install -D tailwindcss@^3 @tailwindcss/forms postcss autoprefixer vite laravel-vite-plugin

These are the correct and supported packages for Laravel + Vite +
TailwindCSS integration.

Do not import or use @tailwindcss/vite and we must remove it from
vite.config.js if it is there.

In our case we have set up postcss.config.js and if it is not, run npx:

$ sail npx tailwindcss init -p

Now we can start vite without error.

$ sail npm run dev