Use firstOrCreate() to fix duplicate seeding
The use of factory()->create() does seed the database, but it has a bug
of duplicate seeding.  To fix it, use User::firstOrCreate() call, which
tries to find the first record that matches the given attribute (e.g.,
email), and if not found, create a new one by using any additional
attributes.

By the way firstOrCreate() is a method on the Eloquent model, and it is
called directly on the model.

For all the default values provided by the factory, use the definition()
call.  For email and name attributes, override them with custom values.

In the current UserFactory, the email is defined as:

  'email' => fake()->unique()->safeEmail(),

Which, returned by User::factory()->definition(), is included in the
definition.

By explicitly including the email in array_merge, it makes certain the
correct email is inserted — even if the factory changes later.

This is great defensive programming by the clear intent in code:

"We want this exact email and want it once only!"