Build username in factory and seeder
This is the most natural and readable way to generate usernames.

  'username' => fake()->userName(),

To avoid duplicate usernames during seeding, use unique() method.
Laravel will keep generating until it finds a unique one.

With username field built in the migration and factory, we also need to
automate seeding process by cleanly integrating the new username field
into the seeder.  Thus, it will be flawless to run this.

$ sail artisan migrate:fresh --seed