Update mass assignment protection
When we add 'username' to the registration logic like in the Volt
component, Laravel would throw an error because 'username' wasn't listed
in $fillable.  By adding 'username' to $fillable, it tells Laravel it's
okay to mass-assign the username field, which allows this line in the
Volt component to work:

  User::create($validated);

Where $validated includes 'username'.