Use both name and username
The use of name and username is not mutually exclusive.  It's better to
use both at the same time.

x-data="{ username: 'janedoe' }"

This sets up an Alpine.js component with a reactive username property.

x-text="username"

This tells Alpine to display the value of username inside <span> tag.

auth()->user()->username

This tells Laravel to fetch the current user's username at render time.

If the logged-in user has both username and name as follows:

username = 'janedoe'
name = 'Jane Doe'

Then this will render:

<span x-data="..." x-text="username">janedoe</span> &nbsp;(Jane Doe)

Which visually appears as:

janedoe (Jane Doe)