Install the core of PestPHP
The pestphp/pest is the main Pest package and it gives the syntax like:

test('it works', function () {
    expect(true)->toBeTrue();
});

And it is required for any PestPHP test.

The pestphp/pest-plugin-laravel adds Laravel-specific features:

1) $this->get(), $this->post(), etc.
2) Laravel assertions like assertStatus(), assertSee()
3) Support for Laravel's testing helpers

This plugin is highly recommended if we're testing a Laravel app and is
required if we're testing controllers, routes, middleware, etc.

Thus, install the core of PestPHP with --dev flag under the require-dev
section of the composer.json file.

$ sail composer require --update-with-all-dependencies --dev -- pestphp/pest pestphp/pest-plugin-laravel

For optional Pest plugins, we can safely install them at any time.