Use dynamic username for the profile
Use the actual username from the database.  This makes the profile
personalized and relevant to the user being viewed.

If we later want to display other users' profiles, using the username
from the database makes that easy and flexible.

When we change the username in the database, it will automatically
appear in the UI without the need to manually update Blade templates.

Fetch the user from the database and pass it to the view.  Apply the
username if available, otherwise show 'Anonymous'.

Use Route::resource() to set the /home/{user} URL for a user profile,
which works well with the controller's show($user) method.  Apply auth
via middleware to be consistent with the rest of the routes.