Use the logged-in user's profile
Use the logged-in user's profile to show the /home page.  The use of
index() method is now safe and purposeful.

Auth::user() is never null due to the auth middleware applied.  No need
for fallback like ?? 'Anonymous' since the user always exists.

The use of show() method is now safe and user-friendly.  If the user
doesn't exist, Laravel will return 404, clean and user-friendly.  If it
gets to the view, no need for fallback as $user exists.

We're using the same view (home.blade.php) for both methods and it is
fine as long as both views look the same.

But if we later want to show different content (e.g., edit button only
on our own profile), we can split the views.