Build ProfileFactory factory
Build the factory with title, description and url so that they will seed
the respective data according to the migration schema.

More importantly we must add user_id foreign key in the Profile model.
This shows the Profile id belongs to the User id.  To set up the value
of user_id, it is the best to use User::factory().

User::factory() means: "If no user exists, create one.  If we're reusing
this in a test with existing users, it can link to one."