Build the migration for Profile
The use of foreignId('user_id')->constrained() creates a foreign key
that references users table id.

The use of ->onDelete('cascade') ensures when a user is deleted, their
profile is deleted as well.

All other fields are nullable, so they don't have to be filled.  But
user_id field is required and the factory must provide a user_id.