Make seeders idempotent
Before the change, get all users regardless of whether they already have
a profile and create a new profile for each one.  If a user already has
a profile, this will create a second (duplicate) profile.  That could
break the app if we assume one profile per user.

After the change, get the users without a profile and prevent duplicate
profiles.  It's safe to run multiple times as it is idempotent.

An operation is idempotent if running it multiple times yields the same
result as running it once.

In Laravel seeding, this is very important, especially when using:

$ sail artisan migrate:fresh --seed

This is safe.  Drop DB, re-seed → no duplicates.

$ sail artisan db:seed

But when we run this, User::all() creates duplicates while it is safe to
run User::doesntHave('profile') as it doesn't create duplicates.

When run User::doesntHave(), make sure User model has this relationship:

// App/Models/User.php
public function profile()
{
    return $this->hasOne(Profile::class);
}

Otherwise, User::doesntHave() won't work.