Remove idle $profile variable
Laravel follows the principle of explicit data passing.  We should pass
the minimal set of data needed, organized in a logical, meaningful way.

The profile belongs to the user and it is accessed via a $user->profile
relation.  It does not stand alone in this context.  Thus, passing of
$user, but not $profile, is sufficient.

We are not hiding anything and we are not making the view work harder.
Eloquent relations are designed exactly for this kind of access.